feat: overhaul of dotfiles and emacs improvements!  (#1)
* fix: zsh files to provide necessary environment settings on startup

* fix: symlinks if folder changes location & dont source zshrc in install

* feat: add prompt for install on each extension

move installation /setup of most packages inside their respective
extension. This should be beneficial long term since most of the
extensions depend on tools that are not OS dependent

* refact: remove unneded package installs & update brew install

* refact: keybindings to emacs init.el section

* feat: add env vars for rust & bump emacs version

* fix: install auto-compile via straight & fonts naming

Ensure to use built-in which-key

* fix: install Docker Desktop for brew since needed for Mac

* update: frozen versions of packages

* fix: installed tree-sitter grammer configs

* update: .gitignore

* fix: shut off bad keymaps for consult for now

* refact: reduce use of utils.sh

* feat: vterm goodies for emacs

* refact: default keybindings in emacs `use-package` setup

* feat: replace `combobulate` with `expand-region`

* feat: vterm goodies for emacs

* fix: git-commit-mode -> git-commit-setup

* remove: `init-keybindings.el`

* fix: keybinding for `projectile`

* refact: keybindings

* fix: `consult-dir` & add `aaronzinhoo--consult-ripgrep-dwim`

`aaronzinhoo--consult-ripgrep-dwim` will search for what is in region
otherwise will open up an empty search.

* feat: add multi completion support for `multiple-cursors`

Really rough implementation of a first attempt at multi completion
support in `multiple-cursors` that utilizes no external dependencies but
introduces a single variable to track the completion candidate across
buffers in a very opinionated manner.

* feat: update indent command for python

* fix: `markdown-mode` code block insert

* refact: minor cleanup to hydras and packages

* feat: try `kubed`!

* fix: orderless completion styles

* feat: add `previous-window-any-frame` to M-; for quick swaps

* fix: `lsp` completion to use orderless

* feat: remove ansible support

* refact: simplify directory structure

* feat: add `install_bootstrap_check` & color prompts

Added `install_bootstrap_check` to reduce changes needed across all
bootstrap extension files. Ran into a bug that needed fixing in a files
and it was a pain 😒.

* refact: standardize prompt structure

May want to split `install_packages` down some more since the specific
package manager cases are essentially identical.

* refact: renamed rust bootstrap file

* feat: slowly replace projectile with project.el

Added `project-vterm` based on `project-shell`. Minor compilation
adjustments.

* fix: remove `delete-pair` key for now to avoid collision

* fix: ensure C-; always runs `evilnc-comment-or-uncomment-lines`

* refact: folder/file names

* update: README.md

* refact: folder/file names and remove `pair-navigation` package

+++ bootstrap_extensions/zsh_bootstrap.sh
@@ -1,12 +1,13 @@
 . "$( pwd )/utils.sh"
 
 PROMPT="[ ZSHExtensionLoader ]: "
-PACKAGE_TO_CHECK="git"
 
-check_system_package_installed $PACKAGE_TO_CHECK
-if [ $? -eq 1 ];then
+install_bootstrap_check "[ -d $HOME/.oh-my-zsh ]" "rust" || exit 0;
+
+PACKAGE_TO_CHECK="git"
+if ! command -v "$PACKAGE_TO_CHECK" > /dev/null; then
     echo_with_prompt "[FAIL] ${PACKAGE_TO_CHECK} not installed"
-    return 1
+    exit 1
 fi
 
 CUSTOM_PLUGINS="$HOME/.oh-my-zsh/custom/plugins"
@@ -16,30 +17,32 @@ sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/too
 
 echo_with_prompt "cloning plugins and themes for OMZ and ZSH"
 
-git clone --depth=1 https://github.com/romkatv/powerlevel10k.git ${ZSH_CUSTOM:-$HOME/.oh-my-zsh/custom}/themes/powerlevel10k
+git clone --depth=1 https://github.com/romkatv/powerlevel10k.git "${ZSH_CUSTOM:-$HOME/.oh-my-zsh/custom}/themes/powerlevel10k"
 
-git clone https://github.com/zsh-users/zsh-autosuggestions "${CUSTOM_PLUGINS}"/zsh-autosuggestions
+git clone https://github.com/zsh-users/zsh-autosuggestions "${CUSTOM_PLUGINS}/zsh-autosuggestions"
 
-git clone https://github.com/mattberther/zsh-pyenv "${CUSTOM_PLUGINS}"/zsh-pyenv
+git clone https://github.com/mattberther/zsh-pyenv "${CUSTOM_PLUGINS}/zsh-pyenv"
 
-git clone https://github.com/zsh-users/zsh-completions "${CUSTOM_PLUGINS}"/zsh-completions
+git clone https://github.com/zsh-users/zsh-completions "${CUSTOM_PLUGINS}/zsh-completions"
 
 git clone https://github.com/zsh-users/zsh-syntax-highlighting.git "${CUSTOM_PLUGINS}/zsh-syntax-highlighting"
 
-git clone https://github.com/agkozak/zsh-z "${CUSTOM_PLUGINS}"/zsh-z
+git clone https://github.com/agkozak/zsh-z "${CUSTOM_PLUGINS}/zsh-z"
 
-git clone https://github.com/TamCore/autoupdate-oh-my-zsh-plugins "${CUSTOM_PLUGINS}"/autoupdate
+git clone https://github.com/TamCore/autoupdate-oh-my-zsh-plugins "${CUSTOM_PLUGINS}/autoupdate"
 
-git clone https://github.com/lukechilds/zsh-nvm "${CUSTOM_PLUGINS}"/zsh-nvm
+git clone https://github.com/lukechilds/zsh-nvm "${CUSTOM_PLUGINS}/zsh-nvm"
 
-git clone https://github.com/lukechilds/zsh-better-npm-completion "${CUSTOM_PLUGINS}"/zsh-better-npm-completion
+git clone https://github.com/lukechilds/zsh-better-npm-completion "${CUSTOM_PLUGINS}/zsh-better-npm-completion"
 
-git clone https://github.com/johanhaleby/kubetail.git "${CUSTOM_PLUGINS}"/kubetail
+git clone https://github.com/johanhaleby/kubetail.git "${CUSTOM_PLUGINS}/kubetail"
 
 
-echo_with_prompt "adding necessary symlinks"
-# link batcat to bat on local
-ln -s /usr/bin/batcat $HOME/.local/bin/bat
+echo_with_prompt "Adding necessary symlinks"
+if [ ! -L "$HOME/.local/bin/bat" ]; then
+    echo_with_prompt "linking batcat"
+    ln -s /usr/bin/batcat $HOME/.local/bin/bat
+fi
 
 if [ -n "$WSL_DISTRO_NAME" ]; then
     echo_with_prompt "changing shell to ZSH lovliness"
